Fix duplicate model classes
