--- app/src/main/java/com/team7/cmput301/android/theirisproject/BodyLocation.java
@@ -1,22 +1,0 @@
-/*
- * Copyright (c) Team 7, CMPUT301, University of Alberta - All Rights Reserved. You may use, distribute, or modify this code under terms and conditions of the Code of Students Behavior at University of Alberta
- */
-
-package com.team7.cmput301.android.theirisproject;
-
-public class BodyLocation implements Location {
-    @Override
-    public double[] asDouble() {
-        return new double[2];
-    }
-
-    @Override
-    public double getX() {
-        return 0;
-    }
-
-    @Override
-    public double getY() {
-        return 0;
-    }
-}
