--- app/src/main/java/com/team7/cmput301/android/theirisproject/Location.java
@@ -1,13 +1,0 @@
-/*
- * Copyright (c) Team 7, CMPUT301, University of Alberta - All Rights Reserved. You may use, distribute, or modify this code under terms and conditions of the Code of Students Behavior at University of Alberta
- */
-
-package com.team7.cmput301.android.theirisproject;
-
-public interface Location {
-    public double[] asDouble();
-
-    public double getX();
-
-    public double getY();
-}
