--- app/src/main/java/com/team7/cmput301/android/theirisproject/PatientList.java
@@ -1,23 +1,0 @@
-/*
- * Copyright (c) Team 7, CMPUT301, University of Alberta - All Rights Reserved. You may use, distribute, or modify this code under terms and conditions of the Code of Students Behavior at University of Alberta
- */
-
-package com.team7.cmput301.android.theirisproject;
-
-import java.util.ArrayList;
-
-public class PatientList {
-    private ArrayList<Patient> patients = new ArrayList<Patient>();
-
-    public ArrayList<Patient> getPatients() {
-        return this.patients;
-    }
-
-    public void addPatient(Patient patient) {
-
-    }
-
-    public void removePatient(Patient patient) {
-
-    }
-}
