--- app/src/main/java/com/team7/cmput301/android/theirisproject/ProblemList.java
@@ -1,35 +1,0 @@
-/*
- * Copyright (c) Team 7, CMPUT301, University of Alberta - All Rights Reserved. You may use, distribute, or modify this code under terms and conditions of the Code of Students Behavior at University of Alberta
- */
-
-package com.team7.cmput301.android.theirisproject;
-
-import java.util.ArrayList;
-
-public class ProblemList {
-    private ArrayList<Problem> problems = new ArrayList<Problem>();
-
-    public ArrayList<Problem> getProblems() {
-        return this.problems;
-    }
-
-    public void ProblemList() {
-
-    }
-
-    public void add(Problem problem) {
-
-    }
-
-    public void delete(Problem problem) {
-
-    }
-
-    public boolean contains(Problem problem) {
-        return false;
-    }
-
-    public int length() {
-        return 0;
-    }
-}
