--- app/src/main/java/com/team7/cmput301/android/theirisproject/Profile.java
@@ -1,37 +1,0 @@
-/*
- * Copyright (c) Team 7, CMPUT301, University of Alberta - All Rights Reserved. You may use, distribute, or modify this code under terms and conditions of the Code of Students Behavior at University of Alberta
- */
-
-package com.team7.cmput301.android.theirisproject;
-
-public class Profile {
-    private String username;
-    private String email;
-    private String phoneNumber;
-
-    public Profile(String username, String email, String phoneNumber) {
-        this.username = username;
-        this.email = email;
-        this.phoneNumber = phoneNumber;
-    }
-
-    public String getUsername() {
-        return username;
-    }
-
-    public String getEmail() {
-        return email;
-    }
-
-    public String getPhoneNumber() {
-        return phoneNumber;
-    }
-
-    public void updateProfile(String name, String email, String phoneNumber) {
-
-    }
-
-    public Profile getProfile() {
-        return null;
-    }
-}
