--- app/src/main/java/com/team7/cmput301/android/theirisproject/RecordPhoto.java
@@ -1,19 +1,0 @@
-/*
- * Copyright (c) Team 7, CMPUT301, University of Alberta - All Rights Reserved. You may use, distribute, or modify this code under terms and conditions of the Code of Students Behavior at University of Alberta
- */
-
-package com.team7.cmput301.android.theirisproject;
-
-import android.graphics.Bitmap;
-
-public class RecordPhoto {
-    private Bitmap img;
-
-    public Bitmap getImg() {
-        return img;
-    }
-
-    public void RecordPhoto(Bitmap photo) {
-
-    }
-}
