+++ app/src/main/java/com/team7/cmput301/android/theirisproject/model/CareProvider.java
@@ -1,15 +1,16 @@
 /*
- * Copyright (c) Team X, CMPUT301, University of Alberta - All Rights Reserved. You may use, distribute, or modify this code under terms and conditions of the Code of Students Behavior at University of Alberta
+ * Copyright (c) Team 7, CMPUT301, University of Alberta - All Rights Reserved. You may use, distribute, or modify this code under terms and conditions of the Code of Students Behavior at University of Alberta
  */
 
 package com.team7.cmput301.android.theirisproject.model;
 
 import java.util.ArrayList;
+import java.util.List;
 
 public class CareProvider extends User {
-    private ArrayList<Patient> patients = new ArrayList<Patient>();
+    private List<Patient> patients = new ArrayList<>();
 
-    public ArrayList<Patient> getPatients() {
+    public List<Patient> getPatients() {
         return this.patients;
     }
 
@@ -17,7 +18,7 @@ public class CareProvider extends User {
         return null;
     }
 
-    public void CareProvider(String name, String email, String phoneNumber) {
-
+    public CareProvider(String name, String email, String phoneNumber) {
+        super("CareProvider", name, email, phoneNumber);
     }
 }
