+++ app/src/main/java/com/team7/cmput301/android/theirisproject/model/Patient.java
@@ -1,16 +1,17 @@
 /*
- * Copyright (c) Team X, CMPUT301, University of Alberta - All Rights Reserved. You may use, distribute, or modify this code under terms and conditions of the Code of Students Behavior at University of Alberta
+ * Copyright (c) Team 7, CMPUT301, University of Alberta - All Rights Reserved. You may use, distribute, or modify this code under terms and conditions of the Code of Students Behavior at University of Alberta
  */
 
 package com.team7.cmput301.android.theirisproject.model;
 
 import java.util.ArrayList;
+import java.util.List;
 
 public class Patient extends User {
-    private ArrayList<CareProvider> careProviders = new ArrayList<CareProvider>();
+    private List<CareProvider> careProviders = new ArrayList<>();
     private ProblemList problems;
 
-    public ArrayList<CareProvider> getCareProviders() {
+    public List<CareProvider> getCareProviders() {
         return this.careProviders;
     }
 
@@ -18,7 +19,7 @@ public class Patient extends User {
         return this.problems;
     }
 
-    public void Patient(String name, String email, String phoneNumber) {
-
+    public Patient(String name, String email, String phoneNumber) {
+        super("Patient", name, email, phoneNumber);
     }
 }
