+++ app/src/main/java/com/team7/cmput301/android/theirisproject/model/Profile.java
@@ -1,5 +1,5 @@
 /*
- * Copyright (c) Team X, CMPUT301, University of Alberta - All Rights Reserved. You may use, distribute, or modify this code under terms and conditions of the Code of Students Behavior at University of Alberta
+ * Copyright (c) Team 7, CMPUT301, University of Alberta - All Rights Reserved. You may use, distribute, or modify this code under terms and conditions of the Code of Students Behavior at University of Alberta
  */
 
 package com.team7.cmput301.android.theirisproject.model;
@@ -8,6 +8,12 @@ public class Profile {
     private String username;
     private String email;
     private String phoneNumber;
+
+    public Profile(String username, String email, String phoneNumber) {
+        this.username = username;
+        this.email = email;
+        this.phoneNumber = phoneNumber;
+    }
 
     public String getUsername() {
         return username;
@@ -19,10 +25,6 @@ public class Profile {
 
     public String getPhoneNumber() {
         return phoneNumber;
-    }
-
-    public void Profile(String name, String email, String phoneNumber) {
-
     }
 
     public void updateProfile(String name, String email, String phoneNumber) {
