+++ app/src/main/java/com/team7/cmput301/android/theirisproject/model/RecordList.java
@@ -4,9 +4,12 @@
 
 package com.team7.cmput301.android.theirisproject.model;
 
-import java.util.ArrayList;
+import android.support.annotation.NonNull;
 
-public class RecordList {
+import java.util.ArrayList;
+import java.util.Iterator;
+
+public class RecordList implements Iterable<Record> {
     private ArrayList<Record> records = new ArrayList<Record>();
 
     public ArrayList<Record> getRecords() {
@@ -31,5 +34,11 @@ public class RecordList {
 
     public int length() {
         return 0;
+    }
+
+    @NonNull
+    @Override
+    public Iterator<Record> iterator() {
+        return records.iterator();
     }
 }
